requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
bitcoin==1.1.42
cached-property==1.5.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
click @ file:///tmp/build/80754af9/click_1646056591685/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==3.2.0
eth-utils==2.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-simplelogin==0.1.1
Flask-SQLAlchemy==3.0.1
Flask-WTF==0.15.1
fonttools==4.25.0
frozenlist==1.3.1
future==0.18.2
greenlet==1.1.3.post0
hexbytes==0.3.0
idna==3.3
ImageHash==4.3.1
imageio==2.22.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipfshttpclient==0.7.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib @ file:///croot/joblib_1666298844297/work
jsonschema==4.15.0
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
lru-dict==1.1.8
lshashpy3==0.0.8
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///opt/conda/conda-bld/matplotlib-suite_1660167928326/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multiaddr==0.0.9
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
mysqlclient==2.1.1
netaddr==0.8.0
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
numpy-quaternion==2022.4.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.1.0.25
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
parsimonious==0.8.1
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
ply==3.11
protobuf==3.20.1
py-ecc==1.4.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.15.0
PyMySQL @ file:///tmp/build/80754af9/pymysql_1610482900657/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.9.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pysha3==1.0.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyWavelets==1.3.0
QFTPy==0.1
requests==2.28.1
rlp==2.0.1
scikit-image==0.19.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
sip==4.19.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==1.4.41
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2021.11.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz==0.12.0
torch==1.13.0
torchvision @ file:///croot/torchvision_1665558889928/work
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.12
varint==1.0.2
web3==5.30.0
websockets==9.1
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.1
